+++ loopslayers/generators/maxpool.hpp
@@ -41,7 +41,7 @@ class MaxpoolGenerator
 public:
     MaxpoolGenerator(Context aCTX) : CTX(aCTX), m_done(false), lanes(CTX.vlanes<_Tp>()) {}
     typename MPGenTraits<_Tp>::maxpool_t generate(int kh_, int kw_, int padding_top, int padding_left, int padding_bottom, int padding_right, int stride_y, int stride_x, int activation_type, float alpha);
-    dwc_algs_limits calc_maxpool_algs_limits(int NC, int H, int W, int kh, int kw, int64_t H0, int64_t W0, int padding_top, int padding_left, int padding_bottom, int padding_right, int stride_y, int stride_x);
+    dwc_algs_limits calc_maxpool_algs_limits(int NC, int H, int W, int kh, int kw, int H0, int W0, int padding_top, int padding_left, int padding_bottom, int padding_right, int stride_y, int stride_x);
 private:
     Context CTX;
     bool m_done;
@@ -518,7 +518,7 @@ typename MPGenTraits<_Tp>::maxpool_t MaxpoolGenerator<_Tp>::generate(int kh_, in
 }
 
 template<typename _Tp>
-dwc_algs_limits MaxpoolGenerator<_Tp>::calc_maxpool_algs_limits(int NC, int H, int W, int kh, int kw, int64_t H0, int64_t W0, int padding_top, int padding_left, int /*padding_bottom*/, int /*padding_right*/, int stride_y, int stride_x)
+dwc_algs_limits MaxpoolGenerator<_Tp>::calc_maxpool_algs_limits(int NC, int H, int W, int kh, int kw, int H0, int W0, int padding_top, int padding_left, int /*padding_bottom*/, int /*padding_right*/, int stride_y, int stride_x)
 {
     int Cms, Cme;
     int lsimd = stride_x == 1 ? upMultipleOf(kw + lanes - 2, lanes) - 1:
